feat(builder-tools,common,connect): use lazy loading on components inside editor
