+++ apps/connect/package.copy.json
@@ -1,7 +1,7 @@
 {
   "name": "@powerhousedao/connect",
   "productName": "Powerhouse-Connect",
-  "version": "4.1.0-dev.105",
+  "version": "4.1.0-dev.109",
   "description": "Powerhouse Connect",
   "main": "dist/index.html",
   "type": "module",
@@ -22,6 +22,11 @@
     "./package.copy.json"
   ],
   "exports": {
+    ".": {
+      "source": "./src/index.ts",
+      "import": "./lib/src/index.js",
+      "types": "./lib/src/index.d.ts"
+    },
     "./package.json": "./package.copy.json",
     "./main.js": {
       "source": "./src/main.tsx",
