+++ apps/connect/src/components/cookie-banner.tsx
@@ -1,4 +1,7 @@
-import { useAcceptedCookies, useCookieBanner } from "@powerhousedao/connect/hooks";
+import {
+  useAcceptedCookies,
+  useCookieBanner,
+} from "@powerhousedao/connect/hooks";
 import type { CookieInput } from "@powerhousedao/design-system/connect";
 import { CookieBanner as PHCookieBanner } from "@powerhousedao/design-system/connect";
 import { showPHModal } from "@powerhousedao/reactor-browser";
