+++ apps/connect/src/components/login.tsx
@@ -3,8 +3,8 @@ import { DotsLoader } from "@powerhousedao/connect/components";
 import { PowerhouseButton } from "@powerhousedao/design-system";
 import { openRenown } from "@powerhousedao/reactor-browser";
 import {
-    useLoginStatus,
-    useUser,
+  useLoginStatus,
+  useUser,
 } from "@powerhousedao/reactor-browser/connect";
 
 export const Login: React.FC = () => {
