+++ apps/connect/src/pages/content.tsx
@@ -1,23 +1,26 @@
-import { DriveEditorContainer, DriveIcon } from "@powerhousedao/connect/components";
 import {
-    connectConfig,
-    defaultPHDocumentEditorConfig,
-    defaultPHDriveEditorConfig,
+  DriveEditorContainer,
+  DriveIcon,
+} from "@powerhousedao/connect/components";
+import {
+  connectConfig,
+  defaultPHDocumentEditorConfig,
+  defaultPHDriveEditorConfig,
 } from "@powerhousedao/connect/config";
 import {
-    HomeScreen,
-    HomeScreenAddDriveItem,
-    HomeScreenItem,
+  HomeScreen,
+  HomeScreenAddDriveItem,
+  HomeScreenItem,
 } from "@powerhousedao/design-system/connect";
 import {
-    setPHDocumentEditorConfig,
-    setPHDriveEditorConfig,
-    setSelectedDrive,
-    useDriveEditorModuleById,
-    useDrives,
-    useSelectedDocument,
-    useSelectedDriveSafe,
-    useSelectedFolder,
+  setPHDocumentEditorConfig,
+  setPHDriveEditorConfig,
+  setSelectedDrive,
+  useDriveEditorModuleById,
+  useDrives,
+  useSelectedDocument,
+  useSelectedDriveSafe,
+  useSelectedFolder,
 } from "@powerhousedao/reactor-browser";
 import type { DocumentDriveDocument } from "document-drive";
 import { useEffect } from "react";
