+++ apps/connect/src/store/editor.ts
@@ -1,13 +1,14 @@
-import { documentModelEditorModule } from "@powerhousedao/builder-tools/editor";
 import "@powerhousedao/builder-tools/style.css";
-import { GenericDriveExplorer } from "@powerhousedao/common/generic-drive-explorer";
+import DriveExplorerEditor from "@powerhousedao/common/generic-drive-explorer/editor-component";
+import DocumentModelEditor from "@powerhousedao/builder-tools/editor-component";
 import type { VetraEditorModule } from "@powerhousedao/reactor-browser";
 import { DEFAULT_DRIVE_EDITOR_ID } from "@powerhousedao/reactor-browser";
+import { lazy } from "react";
 
 export function loadGenericDriveExplorerEditorModule(): VetraEditorModule {
   const name = "Generic Drive Explorer";
-  const documentTypes = GenericDriveExplorer.documentTypes;
-  const Component = GenericDriveExplorer.Component;
+  const documentTypes = ["powerhouse/document-drive"];
+  const Component = DriveExplorerEditor;
   const vetraEditorModule: VetraEditorModule = {
     id: DEFAULT_DRIVE_EDITOR_ID,
     name,
@@ -20,8 +21,8 @@ export function loadGenericDriveExplorerEditorModule(): VetraEditorModule {
 export function loadDocumentModelEditor(): VetraEditorModule {
   const name = "Document Model Editor";
   const id = "document-model-editor-v2";
-  const documentTypes = documentModelEditorModule.documentTypes;
-  const Component = documentModelEditorModule.Component;
+  const documentTypes = ["powerhouse/document-model"];
+  const Component = DocumentModelEditor;
   const vetraEditorModule: VetraEditorModule = {
     id,
     name,
