+++ apps/connect/vite.config.ts
@@ -13,14 +13,17 @@ export default defineConfig(({ mode }) => {
 
   const additionalViteConfig: UserConfig = {
     // add your own vite config here
-    resolve: {
-      conditions: ["source", "browser", "module", "jsnext:main", "jsnext"],
-    },
     plugins: [process.env.PH_DEBUG_BUILD === "true" ? analyzer() : undefined],
     build: {
+      sourcemap: false,
+      minify: true,
       rollupOptions: {
         output: {
           manualChunks: {
+            react: ["react"],
+            "react-dom": ["react-dom"],
+            "react-dom/client": ["react-dom/client"],
+            "react/jsx-runtime": ["react/jsx-runtime"],
             sentry_browser: ["@sentry/browser"],
             sentry_react: ["@sentry/react"],
             graphql_request: ["graphql-request"],
@@ -29,13 +32,16 @@ export default defineConfig(({ mode }) => {
             document_model: ["document-model"],
             document_drive: ["document-drive"],
             drive_explorer: ["@powerhousedao/common/generic-drive-explorer"],
-            document_model_editor: ["@powerhousedao/builder-tools/editor"],
+            document_model_editor: [
+              "@powerhousedao/builder-tools/editor-component",
+            ],
             reactor_browser: ["@powerhousedao/reactor-browser"],
-            common: ["@powerhousedao/common"],
             config: ["@powerhousedao/config"],
             design_system: ["@powerhousedao/design-system"],
             design_system_connect: ["@powerhousedao/design-system/connect"],
+            design_system_ui: ["@powerhousedao/design-system/ui"],
             document_engineering: ["@powerhousedao/document-engineering"],
+            tailwind_merge: ["tailwind-merge"],
           },
         },
       },
