+++ packages/builder-tools/document-model-editor/components/operation.tsx
@@ -1,12 +1,11 @@
 import type { ModuleSpecification } from "document-model";
-import { useCallback } from "react";
+import { lazy, useCallback } from "react";
 import type { DocumentActionHandlers } from "../types/documents.js";
 import { ensureValidOperationSchemaInputName } from "../utils/linting.js";
-import { GraphqlEditor } from "./code-editors/graphql-editor.js";
 import { OperationDescriptionForm } from "./operation-description-form.js";
 import { OperationErrors } from "./operation-errors.js";
 import { OperationForm } from "./operation-form.js";
-
+const GraphqlEditor = lazy(() => import("./code-editors/graphql-editor.js"));
 export type WrappedHandlers = DocumentActionHandlers & {
   addOperationAndInitialSchema: (
     moduleId: string,
