+++ packages/builder-tools/document-model-editor/editor.tsx
@@ -23,11 +23,8 @@ import {
   setStateSchema,
 } from "document-model";
 import { generateId } from "document-model/core";
-import { useEffect, useRef, useState } from "react";
+import { lazy, useEffect, useRef, useState } from "react";
 import { Divider } from "./components/divider.js";
-import { ModelMetadata } from "./components/model-metadata-form.js";
-import { Modules } from "./components/modules.js";
-import { StateSchemas } from "./components/state-schemas.js";
 import { editorConfig } from "./config.js";
 import { SchemaContextProvider } from "./context/schema-context.js";
 import { useSelectedDocumentModelDocument } from "./hooks/useDocumentModelDocument.js";
@@ -37,6 +34,9 @@ import {
   initializeModelSchema,
   makeOperationInitialDoc,
 } from "./utils/helpers.js";
+import ModelMetadata from "./components/model-metadata-form.js";
+import Modules from "./components/modules.js";
+const StateSchemas = lazy(() => import("./components/state-schemas.js"));
 
 export default function Editor() {
   useSetPHDocumentEditorConfig(editorConfig);
